requirements.txt
fastapi
uvicorn[standard]
requests
pathspec
openai
pydantic
python-dotenv